+++ HTTPShortcuts/app/src/main/kotlin/ch/rmy/android/http_shortcuts/activities/BaseFragment.kt
@@ -39,4 +39,6 @@ open class BaseFragment : Fragment() {
 
     }
 
+    open fun onBackPressed(): Boolean = false
+
 }